innhabitserver/requirements.txt
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
certifi==2025.1.31
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cookiecutter==2.6.0
distlib==0.3.9
Django==5.1.7
django-browser-reload==1.18.0
django-environ==0.12.0
django-htmx==1.23.0
django-tailwind==3.8.0
django-template-partials==24.4
django_components==0.134
djangorestframework==3.15.2
djc_core_html_parser==1.0.2
filelock==3.18.0
gunicorn==23.0.0
identify==2.6.9
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mypy==1.15.0
mypy-extensions==1.0.0
nodeenv==1.9.1
numpy==2.2.4
packaging==24.2
pillow==11.1.0
platformdirs==4.3.7
pre_commit==4.2.0
psycopg2-binary==2.9.10
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
sentry-sdk==2.24.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
types-python-dateutil==2.9.0.20241206
types-requests==2.32.0.20250306
types-six==1.17.0.20250304
typing_extensions==4.12.2
urllib3==2.3.0
virtualenv==20.29.3
